requirements.txt
torch~=1.13.1
PyYAML~=6.0
Pillow~=9.4.0
torchvision~=0.14.1
numpy~=1.24.2
matplotlib~=3.7.1
pandas~=1.5.3
seaborn~=0.12.2
requests~=2.28.2
tqdm~=4.65.0
psutil~=5.9.4
thop~=0.1.1.post2209072238
streamlit~=1.20.0
opencv-python-headless>=4.1.1